Refactoring - consistent interfaces naming
