--- src/common/com/intellij/plugins/haxe/model/IHaxeEnumModel.java
@@ -1,29 +1,0 @@
-/*
- * Copyright 2018 Ilya Malanin
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.intellij.plugins.haxe.model;
-
-import org.jetbrains.annotations.NotNull;
-
-import java.util.List;
-import java.util.stream.Stream;
-
-public interface IHaxeEnumModel {
-  HaxeEnumValueModel getValue(@NotNull String name);
-
-  List<HaxeEnumValueModel> getValues();
-
-  Stream<HaxeEnumValueModel> getValuesStream();
-}
